requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
alembic==1.13.3
pydantic==2.9.2
python-dotenv==1.0.1
jinja2==3.1.4
twilio==9.3.0
httpx==0.27.2
